refactor(bite-common): 重构 bite-common 模块并添加 Redis 支持
- 新增 bite-common-redis 模块，提供 Redis 支持
- 重构 bite-common-core、bite-common-domain等模块的依赖关系
- 更新线程池配置，增加空闲线程存活时间设置
- 删除未使用的数据库初始化脚本- 调整项目结构，优化模块划分
