+++ .idea/compiler.xml
@@ -10,8 +10,8 @@
         <module name="bite-mstemplate-api" />
         <module name="bite-common-core" />
         <module name="bite-common-domain" />
+        <module name="bite-common-redis" />
         <module name="bite-gateway" />
-        <module name="bite-common" />
         <module name="bite-commom-security" />
         <module name="bite-mstemplate-service" />
       </profile>
@@ -21,9 +21,10 @@
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
       <module name="FrameworkJava" options="" />
       <module name="bite-commom-security" options="-parameters" />
-      <module name="bite-common" options="-parameters" />
+      <module name="bite-common" options="" />
       <module name="bite-common-core" options="-parameters" />
       <module name="bite-common-domain" options="-parameters" />
+      <module name="bite-common-redis" options="-parameters" />
       <module name="bite-gateway" options="-parameters" />
       <module name="bite-mstemplate" options="" />
       <module name="bite-mstemplate-api" options="-parameters" />
