+++ bite-common/bite-commom-security/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>com.bitejiuyeke</groupId>
-        <artifactId>FrameworkJava</artifactId>
+        <artifactId>bite-common</artifactId>
         <version>0.0.1-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
