+++ bite-common/pom.xml
@@ -10,6 +10,10 @@
     </parent>
 
     <artifactId>bite-common</artifactId>
+    <packaging>pom</packaging>
+    <modules>
+        <module>bite-common-redis</module>
+    </modules>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
