--- deploy/dev/app/mysql/init/init.sql
@@ -1,2 +1,0 @@
-source /opt/sql/inituser.sql;
-source /opt/sql/nacos.sql;
