--- deploy/dev/app/nacos/conf/application.properties
@@ -1,54 +1,0 @@
-# spring
-server.servlet.contextPath=${SERVER_SERVLET_CONTEXTPATH:/nacos}
-server.contextPath=/nacos
-server.port=${NACOS_APPLICATION_PORT:8848}
-server.tomcat.accesslog.max-days=30
-server.tomcat.accesslog.pattern=%h %l %u %t "%r" %s %b %D %{User-Agent}i %{Request-Source}i
-#日志信息 开发环境开启，后续生产看下性能影响
-server.tomcat.accesslog.enabled=${TOMCAT_ACCESSLOG_ENABLED:false}
-server.error.include-message=ALWAYS
-# default current work dir
-server.tomcat.basedir=/home/bit/nacos/tomcat/logs
-#*************** Config Module Related Configurations ***************#
-### Deprecated configuration property, it is recommended to use `spring.sql.init.platform` replaced.
-#spring.datasource.platform=${SPRING_DATASOURCE_PLATFORM:}
-spring.sql.init.platform=${SPRING_DATASOURCE_PLATFORM:mysql}
-nacos.cmdb.dumpTaskInterval=3600
-nacos.cmdb.eventTaskInterval=10
-nacos.cmdb.labelTaskInterval=300
-nacos.cmdb.loadDataAtStart=false
-db.num=${MYSQL_DATABASE_NUM:1}
-db.url.0=jdbc:mysql://${MYSQL_SERVICE_HOST}:${MYSQL_SERVICE_PORT:3306}/${MYSQL_SERVICE_DB_NAME}?${MYSQL_SERVICE_DB_PARAM:characterEncoding=utf8&connectTimeout=1000&socketTimeout=3000&autoReconnect=true&useSSL=false}
-db.user.0=${MYSQL_SERVICE_USER}
-db.password.0=${MYSQL_SERVICE_PASSWORD}
-### The auth system to use, currently only 'nacos' and 'ldap' is supported:
-nacos.core.auth.system.type=${NACOS_AUTH_SYSTEM_TYPE:nacos}
-### worked when nacos.core.auth.system.type=nacos
-### The token expiration in seconds:
-nacos.core.auth.plugin.nacos.token.expire.seconds=${NACOS_AUTH_TOKEN_EXPIRE_SECONDS:18000}
-### The default token:
-#nacos.core.auth.plugin.nacos.token.secret.key=${NACOS_AUTH_TOKEN:}+
-nacos.core.auth.plugin.nacos.token.secret.key=VGhpc0lzTXlDdXN0b21TZWNyZXRLZXkwMTIzNDU2Nzg=
-### Turn on/off caching of auth information. By turning on this switch, the update of auth information would have a 15 seconds delay.
-nacos.core.auth.enabled=true
-nacos.core.auth.caching.enabled=${NACOS_AUTH_CACHE_ENABLE:false}
-nacos.core.auth.enable.userAgentAuthWhite=${NACOS_AUTH_USER_AGENT_AUTH_WHITE_ENABLE:false}
-nacos.core.auth.server.identity.key=${NACOS_AUTH_IDENTITY_KEY:BITDSSSDKEY1@76186138}
-nacos.core.auth.server.identity.value=${NACOS_AUTH_IDENTITY_VALUE:BITDSSS@45646YUI138}
-## spring security config
-### turn off security
-nacos.security.ignore.urls=${NACOS_SECURITY_IGNORE_URLS:/,/error,/**/*.css,/**/*.js,/**/*.html,/**/*.map,/**/*.svg,/**/*.png,/**/*.ico,/console-fe/public/**,/v1/auth/**,/v1/console/health/**,/actuator/**,/v1/console/server/**}
-# metrics for elastic search
-management.metrics.export.elastic.enabled=false
-management.metrics.export.influx.enabled=false
-nacos.naming.distro.taskDispatchThreadCount=10
-nacos.naming.distro.taskDispatchPeriod=200
-nacos.naming.distro.batchSyncKeyCount=1000
-nacos.naming.distro.initDataRatio=0.9
-nacos.naming.distro.syncRetryDelay=5000
-nacos.naming.data.warmup=true
-nacos.naming.empty-service.auto-clean=true
-nacos.naming.empty-service.clean.initial-delay-ms=50000
-nacos.naming.empty-service.clean.period-time-ms=30000
-management.endpoints.web.exposure.include=*
-nacos.istio.mcp.server.enabled=false
